Move helper functions to core config.
