+++ newrelic/core/stats_engine.py
@@ -41,7 +41,7 @@ from newrelic.core.metric import TimeMetric
 from newrelic.core.stack_trace import exception_stack
 
 from newrelic.api.settings import STRIP_EXCEPTION_MESSAGE
-from newrelic.api.time_trace import is_expected_error
+from newrelic.core.config import is_expected_error
 from newrelic.common.encoding_utils import json_encode
 from newrelic.common.streaming_utils import StreamBuffer
